art_marketplace/requirements.txt
Django==5.0.2
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.60.1
python-dotenv==1.0.1 